--- raven-app/src/components/layout/HeadingBar/index.ts
@@ -1,1 +1,0 @@
-export { HeadingBar } from "./HeadingBar";
