--- raven-app/src/components/layout/ProjectTopbar/index.ts
@@ -1,1 +1,0 @@
-export { ProjectTopbar } from './ProjectTopbar'
