--- raven-app/src/components/layout/TabItem/index.ts
@@ -1,1 +1,0 @@
-export { TabItem, TabItemLabel, TabItemProgress } from './TabItem'
